Custom keyboard.io keymap
The main problem with the default layout is that a number of important
keys are missing from the left side, making it difficult to use it with
games. On top of that, as a user of various national layouts, I totally
need right alt in a much better position. It is a key that I use very
frequently.

Of special note is usage of qukeys, which means that I get the special
behaviour of PgUp and PgDown as Control and Shift when pressed together
with other keys already in firmware, so without any special tricks in
different operating systems.